requirements.txt
opencv-python
numpy
Pillow
pandas
sqlalchemy
pymysql
opencv-contrib-python